reset

+++ src/Illuminate/Queue/DatabaseQueue.php
@@ -6,7 +6,6 @@ use DateTime;
 use Carbon\Carbon;
 use Illuminate\Database\Connection;
 use Illuminate\Queue\Jobs\DatabaseJob;
-use Illuminate\Database\Query\Expression;
 use Illuminate\Contracts\Queue\Queue as QueueContract;
 
 class DatabaseQueue extends Queue implements QueueContract
@@ -160,70 +159,91 @@ class DatabaseQueue extends Queue implements QueueContract
     {
         $queue = $this->getQueue($queue);
 
-        if (! is_null($this->expire)) {
-            $this->releaseJobsThatHaveBeenReservedTooLong($queue);
-        }
+        $this->database->beginTransaction();
 
-        if ($job = $this->reserveNextAvailableJob($queue)) {
+        if ($job = $this->getNextAvailableJob($queue)) {
+            $job = $this->markJobAsReserved($job);
+
+            $this->database->commit();
+
             return new DatabaseJob(
                 $this->container, $this, $job, $queue
             );
         }
+
+        $this->database->commit();
     }
 
     /**
-     * Release the jobs that have been reserved for too long.
-     *
-     * @param  string  $queue
-     * @return void
-     */
-    protected function releaseJobsThatHaveBeenReservedTooLong($queue)
-    {
-        $expired = Carbon::now()->subSeconds($this->expire)->getTimestamp();
-
-        $this->database->table($this->table)
-                    ->where('queue', $this->getQueue($queue))
-                    ->where('reserved', 1)
-                    ->where('reserved_at', '<=', $expired)
-                    ->update([
-                        'attempts' => new Expression('attempts + 1'),
-                        'reservation' => null,
-                        'reserved' => 0,
-                        'reserved_at' => null,
-                    ]);
-    }
-
-    /**
-     * Reserve the next available job for the queue.
+     * Get the next available job for the queue.
      *
      * @param  string|null  $queue
      * @return \StdClass|null
      */
-    protected function reserveNextAvailableJob($queue)
+    protected function getNextAvailableJob($queue)
     {
-        $reservation = str_random(48);
-
-        $count = $this->database->table($this->table)
-                      ->where('queue', $this->getQueue($queue))
-                      ->where('reserved', 0)
-                      ->where('available_at', '<=', $this->getTime())
-                      ->orderBy('id', 'asc')
-                      ->limit(1)
-                      ->update([
-                          'reservation' => $reservation,
-                          'reserved'    => 1,
-                          'reserved_at' => $this->getTime(),
-                      ]);
-
-        if ($count == 0) {
-            return;
-        }
-
         $job = $this->database->table($this->table)
-                    ->where('reservation', $reservation)
+                    ->lockForUpdate()
+                    ->where('queue', $this->getQueue($queue))
+                    ->where(function ($query) {
+                        $this->isAvailable($query);
+                        $this->isReservedButExpired($query);
+                    })
+                    ->orderBy('id', 'asc')
                     ->first();
 
-        return (object) $job;
+        return $job ? (object) $job : null;
+    }
+
+    /**
+     * Modify the query to check for available jobs.
+     *
+     * @param  \Illuminate\Database\Query\Builder  $query
+     * @return void
+     */
+    protected function isAvailable($query)
+    {
+        $query->where(function ($query) {
+            $query->where('reserved', 0);
+            $query->where('available_at', '<=', $this->getTime());
+        });
+    }
+
+    /**
+     * Modify the query to check for jobs that are reserved but have expired.
+     *
+     * @param  \Illuminate\Database\Query\Builder  $query
+     * @return void
+     */
+    protected function isReservedButExpired($query)
+    {
+        $expiration = Carbon::now()->subSeconds($this->expire)->getTimestamp();
+
+        $query->orWhere(function ($query) use ($expiration) {
+            $query->where('reserved', 1);
+            $query->where('reserved_at', '<=', $expiration);
+        });
+    }
+
+    /**
+     * Mark the given job ID as reserved.
+     *
+     * @param \stdClass $job
+     * @return \stdClass
+     */
+    protected function markJobAsReserved($job)
+    {
+        $job->reserved = 1;
+        $job->attempts = $job->attempts + 1;
+        $job->reserved_at = $this->getTime();
+
+        $this->database->table($this->table)->where('id', $job->id)->update([
+            'reserved' => $job->reserved,
+            'reserved_at' => $job->reserved_at,
+            'attempts' => $job->attempts,
+        ]);
+
+        return $job;
     }
 
     /**
@@ -235,7 +255,13 @@ class DatabaseQueue extends Queue implements QueueContract
      */
     public function deleteReserved($queue, $id)
     {
-        $this->database->table($this->table)->where('id', $id)->delete();
+        $this->database->beginTransaction();
+
+        if ($this->database->table($this->table)->lockForUpdate()->find($id)) {
+            $this->database->table($this->table)->where('id', $id)->delete();
+        }
+
+        $this->database->commit();
     }
 
     /**
@@ -264,13 +290,12 @@ class DatabaseQueue extends Queue implements QueueContract
     {
         return [
             'queue' => $queue,
-            'payload' => $payload,
             'attempts' => $attempts,
             'reserved' => 0,
             'reserved_at' => null,
             'available_at' => $availableAt,
             'created_at' => $this->getTime(),
-            'reservation' => null,
+            'payload' => $payload,
         ];
     }
